+++ eio-okta-sync/src/command.rs
@@ -1,6 +1,7 @@
 mod check_version;
 mod current;
 mod generate;
+mod make_archive;
 mod mappings;
 mod members;
 mod snapshot;
@@ -14,6 +15,7 @@ pub enum Command {
   CheckVersion(check_version::Command),
   Current(current::Command),
   Generate(generate::Command),
+  MakeArchive(make_archive::Command),
   Mappings(mappings::Command),
   Members(members::Command),
   Snapshot(snapshot::Command),
@@ -25,6 +27,7 @@ impl Command {
       Self::CheckVersion(command) => command.run(),
       Self::Current(command) => command.run(),
       Self::Generate(command) => command.run(),
+      Self::MakeArchive(command) => command.run(),
       Self::Mappings(command) => command.run(),
       Self::Members(command) => command.run().await,
       Self::Snapshot(command) => command.run(),